♻️ refactor: project list page apply vite
