+++ src/app.tsx
@@ -2,7 +2,8 @@ import { Route, Routes } from "react-router-dom";
 
 import { Header } from "./@shared/Header";
 import { Footer } from "./@shared/Footer";
-import { Landing } from "./landing";
+import { Landing } from "./Landing";
+import { ProjectList } from "./ProjectList";
 import * as css from "./app.css";
 
 export function App() {
@@ -11,6 +12,7 @@ export function App() {
       <Header />
       <Routes>
         <Route path="/" element={<Landing />} />
+        <Route path="/projects" element={<ProjectList />} />
       </Routes>
       <Footer />
     </div>
